flask-be/requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.7
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==3.0.1
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.0
MarkupSafe==2.1.4
mongoengine==0.27.0
pycparser==2.21
pymongo==4.6.1
python-dotenv==1.0.1
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
PyJWT~=2.8.0
pillow~=10.2.0
bcrypt~=4.1.2